refactor(ADR-011): migrar cadena completa de pagos de float a int
RESOLVER INCONSISTENCIA: El frontend usa enteros (CLP) pero el backend
todavía usaba float en la cadena completa de procesamiento de pagos.

CAMBIOS EN CADENA (de request a response):

1. StorePaymentRequest: validación numeric→integer, min:0.01→min:1
2. PaymentController: (float) casts → (int)
3. PaymentService: parámetros float→int, comparación < 0.01 → < 1
4. PaymentMethod: acceptsAmount(float) → acceptsAmount(int)
5. PaymentResource: (float) casts → (int)
6. BillResource: todos los campos monetarios (float) → (int)
7. CashSessionResource: opening/closing/expected (float) → (int)

JUSTIFICACIÓN:
- CLP no tiene decimales (moneda chilena)
- ADR-011: Modelo chileno con IVA incluido y valores enteros
- Elimina errores de precisión con floats (0.1 + 0.2 ≠ 0.3)
- Consistencia frontend ↔ backend

ARCHIVOS:
- StorePaymentRequest.php: validación integer
- PaymentController.php: cast (int)
- PaymentService.php: parámetros y lógica integer
- PaymentMethod.php: acceptsAmount(int)
- PaymentResource.php: response integer
- BillResource.php: response integer
- CashSessionResource.php: response integer

+++ app/Modules/Payments/Domain/Entities/PaymentMethod.php
@@ -65,12 +65,12 @@ class PaymentMethod extends Model
 /**
      * Verifica si un monto es válido para este método.
      */
-    public function acceptsAmount(float $amount): bool
+    public function acceptsAmount(int $amount): bool  // ADR-011: integer CLP
     {
         if ($this->max_amount === null) {
             return true;
         }
-        return $amount <= (float) $this->max_amount;
+        return $amount <= (int) $this->max_amount;  // ADR-011: integer comparison
     }
 
     /**
